move context manager to utils, add test
